# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hacker Dashboard API",
    "description": "임의 투자 데이터 스키마에서 자동 분석 뷰를 생성하는 금융 대시보드 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/market/symbols/search": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Search Symbols",
        "description": "심볼 통합 검색 (upbit + yahoo + naver_kr + alias).\n\n- 4개 소스를 asyncio.gather 로 병렬 호출\n- (market, symbol) 기준 dedupe, 높은 score 유지\n- score 내림차순 정렬 후 상위 50개 반환\n- score 필드는 응답에 포함되지 않음 (SymbolInfo.score 는 exclude=True)",
        "operationId": "search_symbols_market_symbols_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "검색 키워드",
              "title": "Q"
            },
            "description": "검색 키워드"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SymbolInfo"
                  },
                  "title": "Response Search Symbols Market Symbols Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/quotes/{symbol}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Quote",
        "description": "단일 심볼 시세 조회 (stub — legacy 호환)",
        "operationId": "get_quote_market_quotes__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/quotes/{market}/{code}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Market Quote",
        "description": "market별 단일 심볼 현재가 조회. Redis 캐시(TTL 5s) 적용.",
        "operationId": "get_market_quote_market_quotes__market___code__get",
        "parameters": [
          {
            "name": "market",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Market"
            }
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/ohlc/{market}/{code}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Ohlc",
        "description": "OHLC 캔들 조회. Redis 캐시(TTL 60s) 적용.",
        "operationId": "get_ohlc_market_ohlc__market___code__get",
        "parameters": [
          {
            "name": "market",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Market"
            }
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1d|1m)$",
              "default": "1d",
              "title": "Interval"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OhlcBar"
                  },
                  "title": "Response Get Ohlc Market Ohlc  Market   Code  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/watchlist/items": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "List Watchlist",
        "description": "워치리스트 조회 (user_id='demo' 고정).",
        "operationId": "list_watchlist_market_watchlist_items_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItemResponse"
                  },
                  "type": "array",
                  "title": "Response List Watchlist Market Watchlist Items Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Add Watchlist",
        "description": "워치리스트에 심볼 추가.",
        "operationId": "add_watchlist_market_watchlist_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/watchlist/items/{item_id}": {
      "delete": {
        "tags": [
          "market"
        ],
        "summary": "Delete Watchlist",
        "description": "워치리스트 항목 삭제.",
        "operationId": "delete_watchlist_market_watchlist_items__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OhlcBar": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts",
            "description": "ISO-8601 UTC"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "OhlcBar",
        "description": "OHLC 캔들 하나"
      },
...
      "Quote": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "description": "upbit | binance | yahoo | naver_kr",
            "default": "unknown"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "change": {
            "type": "number",
            "title": "Change",
            "description": "전일 대비 변화량",
            "default": 0.0
          },
          "change_pct": {
            "type": "number",
            "title": "Change Pct",
            "description": "전일 대비 변화율(%)",
            "default": 0.0
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "KRW | USD | USDT 등",
            "default": "USD"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO-8601 UTC"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "price",
          "timestamp"
        ],
        "title": "Quote"
      },
...
      "SymbolInfo": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "asset_class": {
            "type": "string",
            "title": "Asset Class"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "asset_class",
          "market"
        ],
        "title": "SymbolInfo",
        "description": "심볼 검색 결과"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchlistItemCreate": {
        "properties": {
          "market": {
            "type": "string",
            "title": "Market",
            "description": "upbit | binance | yahoo | naver_kr"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "심볼/코드 (예: KRW-BTC, AAPL)"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          }
        },
        "type": "object",
        "required": [
          "market",
          "code"
        ],
        "title": "WatchlistItemCreate"
      },
      "WatchlistItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO-8601 UTC"
          }
        },
        "type": "object",
        "required": [
          "id",
          "market",
          "code",
          "created_at"
        ],
        "title": "WatchlistItemResponse"
      }
    }
  }
}